[Ide] Add ABI compat shims into own file
